Commit 2

+++ src/main/java/com/example/bookdemo/dao/BookRepository.java
@@ -3,6 +3,6 @@ package com.example.bookdemo.dao;
 import com.example.bookdemo.entity.Book;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface BookRepository extends JpaRepository<Book, Integer> {
-    Book getBookById(int id);
+public interface BookRepository extends JpaRepository<Book, Long> {
+    Book getBookById(Long id);
 }
